+++ src/components/DatePicker/DateRangePicker.story.js
@@ -46,8 +46,7 @@ stories
       />
       <p>I appear after the calendar.</p>
     </Container>
-  ))
-  )
+  )))
   .add('With absolute positioning', (() => (
     <Container>
       <DateRangePicker
@@ -59,8 +58,7 @@ stories
       />
       <p>I should be behind the calendar when opened</p>
     </Container>
-  ))
-  )
+  )))
   .add('Focused initially', (() => (
     <Container>
       <DateRangePicker
@@ -70,8 +68,7 @@ stories
         startDateInputId='start-date-id-0'
       />
     </Container>
-  ))
-  )
+  )))
   .add('With initial start date', (() => (
     <Container>
       <DateRangePicker
@@ -82,8 +79,7 @@ stories
         startDateInputId='start-date-id-0'
       />
     </Container>
-  ))
-  )
+  )))
   .add('With past dates selectable', (() => (
     <Container>
       <DateRangePicker
@@ -95,8 +91,7 @@ stories
         startDateInputId='start-date-id-0'
       />
     </Container>
-  ))
-  )
+  )))
   .add('Remains open after date selection', (() => (
     <Container>
       <DateRangePicker
@@ -107,8 +102,7 @@ stories
         startDateInputId='start-date-id-0'
       />
     </Container>
-  ))
-  )
+  )))
   .add('Remains open always, even when clicking outside component', (() => (
     <Container>
       <DateRangePicker
@@ -119,8 +113,7 @@ stories
         startDateInputId='start-date-id-0'
       />
     </Container>
-  ))
-  )
+  )))
   .add('With preset options panel', (() => (
     <Container>
       <DateRangePicker
@@ -137,8 +130,7 @@ stories
         startDateInputId='start-date-id-0'
       />
     </Container>
-  ))
-  )
+  )))
   .add('With buttons panel', (() => (
     <Container>
       <DateRangePicker
@@ -154,8 +146,20 @@ stories
         startDateInputId='start-date-id-0'
       />
     </Container>
-  ))
-  )
+  )))
+  .add('With no border on calendar', (() => (
+    <Container>
+      <DateRangePicker
+        endDateInputId='end-date-id-0'
+        focusedInput='startDate'
+        isBorderless={ boolean('isBorderless', true) }
+        isPastDateSelectable={ boolean('isPastDateSelectable', true) }
+        keepOpenOnDateSelect={ boolean('keepOpenOndateSelect', true) }
+        onDatesChange={ action('onDatesChange') }
+        startDateInputId='start-date-id-0'
+      />
+    </Container>
+  )))
   .add('With preset options and no border on calendar', (() => (
     <Container>
       <DateRangePicker
@@ -173,9 +177,8 @@ stories
         startDateInputId='start-date-id-0'
       />
     </Container>
-  ))
-  )
-  .add('With no border on calendar', (() => (
+  )))
+  .add('With panelButtons as an array (not render props)', (() => (
     <Container>
       <DateRangePicker
         endDateInputId='end-date-id-0'
@@ -184,11 +187,48 @@ stories
         isPastDateSelectable={ boolean('isPastDateSelectable', true) }
         keepOpenOnDateSelect={ boolean('keepOpenOndateSelect', true) }
         onDatesChange={ action('onDatesChange') }
+        panelButtons={ [
+          <Button key={ 0 } style="plain" onClick={ action('onClick Clear') }>Clear</Button>,
+          <Button key={ 1 } style="highlight" onClick={ action('onClick Apply') }>Apply</Button>,
+        ] }
+        presetPanelOptions={ presetPanelOptions }
         startDateInputId='start-date-id-0'
       />
     </Container>
-  ))
-  );
+  )))
+  /* eslint-disable react/jsx-no-bind */
+  .add('With panelButtons as render props function returning an array', (() => (
+    <Container>
+      <DateRangePicker
+        endDateInputId='end-date-id-0'
+        focusedInput='startDate'
+        isBorderless={ boolean('isBorderless', true) }
+        isPastDateSelectable={ boolean('isPastDateSelectable', true) }
+        keepOpenOnDateSelect={ boolean('keepOpenOndateSelect', true) }
+        onDatesChange={ action('onDatesChange') }
+        panelButtons={ ({ onDatesChange }) => ([
+          <Button
+            key={ 0 }
+            style="plain"
+            onClick={ () => {
+              action('onClick Clear with onDatesChange via render props')();
+              onDatesChange({ startDate: null, endDate: null });
+            } }>
+            Clear
+          </Button>,
+          <Button
+            key={ 1 }
+            style="highlight"
+            onClick={ action('onClick Apply') }>
+            Apply
+          </Button>,
+        ]) }
+        presetPanelOptions={ presetPanelOptions }
+        startDateInputId='start-date-id-0'
+      />
+    </Container>
+    /* eslint-enable react/jsx-no-bind */
+  )));
 
 const Container = styled.div`
   height: 100vh;